changed senti to float!

+++ frontend/app/run.py
@@ -106,7 +106,7 @@ def create_indexes():
             "properties": {
                 "name": {"type": "text"},
                 "text": {"type": "text"},
-                "senti": {"type": "integer"},
+                "senti": {"type": "float"},
                 "datetime": {"type": "date"},
             }
         }
